refactor(kv-router): split monolithic indexer.rs into indexer/ directory
Move the 4005-line indexer.rs into focused submodules under src/indexer/:
types, metrics, traits, thread_pool, kv_indexer, local, sharded, tests.

Relocate radix_tree, concurrent_radix_tree, nested_map (→ positional),
naive_indexers (→ naive), and approx (→ pruning) into the same directory.

Backward-compat re-exports in lib.rs preserve all public API paths.
