+++ lib/kv-router/src/indexer/mod.rs
@@ -1,0 +1,62 @@
+// SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! KV RadixTree
+//!
+//! This module implements a key-value (KV) store using a Radix Tree structure to efficiently manage and retrieve data blocks.
+//! It is designed to support LLM (Large Language Model) inference by re-using a global KV cache.
+//!
+//! # Overview
+//!
+//! The main components of this module include:
+//!
+//! - **Radix Tree Structure**:
+//!   - The `RadixTree` struct represents the main data structure, with nodes (`RadixBlock`) containing children and associated worker IDs.
+//!   - It allows efficient storage and retrieval of data blocks based on their hashes.
+//!
+//! - **Event Handling**:
+//!   - The `RouterEvent` struct represents events emitted by LLM workers, which can be applied to the Radix Tree to update its state.
+//!   - The `KvIndexer` struct manages these events and match requests asynchronously using Tokio channels.
+//!
+//! - **Hash Computation**:
+//!   - Functions like `compute_block_hash` and `compute_block_hash_for_seq` compute hashes for data blocks and sequences of tokens, facilitating quick lookups.
+//!
+//! - **Concurrency and Asynchronous Operations**:
+//!   - The `KvIndexer` uses a single-threaded Tokio runtime to handle events and match requests concurrently, ensuring efficient processing without blocking.
+//!
+//! - **Match Requests**:
+//!   - The `MatchRequest` struct represents requests to find matches in the Radix Tree, returning overlap scores indicating the best matches.
+//!
+//! # Purpose
+//!
+//! This module provides a scalable and efficient way to manage and retrieve data blocks for LLM inference, leveraging a global KV cache to optimize performance.
+
+mod kv_indexer;
+mod local;
+mod metrics;
+mod sharded;
+mod thread_pool;
+mod traits;
+mod types;
+
+pub mod concurrent_radix_tree;
+#[cfg(feature = "bench")]
+pub mod naive;
+pub mod positional;
+pub mod pruning;
+pub mod radix_tree;
+
+#[cfg(test)]
+mod tests;
+
+// Re-export everything that was public in the old single-file module.
+pub use kv_indexer::*;
+pub use local::*;
+pub use metrics::*;
+pub use sharded::*;
+pub use thread_pool::*;
+pub use traits::*;
+pub use types::*;
+
+// Re-export RadixTree (was `pub use crate::radix_tree::RadixTree` in old indexer.rs)
+pub use radix_tree::RadixTree;
