+++ lib/kv-router/src/lib.rs
@@ -6,19 +6,20 @@
 //! This crate provides the core radix tree implementation and protocols for
 //! efficient KV cache lookup and routing in distributed LLM inference systems.
 
-pub mod approx;
-pub mod concurrent_radix_tree;
 pub mod indexer;
-#[cfg(feature = "bench")]
-pub mod naive_indexers;
-pub mod nested_map;
 pub mod protocols;
-pub mod radix_tree;
 pub mod scheduling;
 pub mod sequences;
 pub mod zmq_wire;
 
-// Backward-compat re-exports: preserve old module paths for external consumers
+// Backward-compat re-exports: old top-level module paths still work
+pub use indexer::concurrent_radix_tree;
+#[cfg(feature = "bench")]
+pub use indexer::naive as naive_indexers;
+pub use indexer::positional as nested_map;
+pub use indexer::pruning as approx;
+pub use indexer::radix_tree;
+
 pub use scheduling::config;
 pub use scheduling::queue;
 pub use scheduling::selector;
